Rewriting Octokit from structs + combine to protocol + async/await
