+++ Packages/Sources/Core/Octokit/Interface/Octokit/Octokit.Client.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public extension Octokit {
+  typealias Client = OctokitClient
+}
+
+public protocol OctokitClient {
+  var loginService: Octokit.LoginService { get }
+  var userService: Octokit.UserService { get }
+}
