--- Packages/Sources/Core/Octokit/Live/Services/LoginService+live.swift
@@ -1,15 +1,0 @@
-import Combine
-import Foundation
-import Octokit
-
-internal extension Octokit.LoginService {
-  static func live(secretsService: SecretsService) -> Octokit.LoginService {
-    Self(
-      login: { [secretsService] token in
-        secretsService
-          .store(.privateAccessToken, value: token)
-          .mapError(PrivateAccessTokenLoginError.unableToStorePrivateAccessToken)
-      }
-    )
-  }
-}
