+++ Packages/Sources/Core/Octokit/Live/api_2022_11_28/Internals/NetworkClient/NetworkClientError.swift
