+++ Packages/Sources/Core/Octokit/Live/api_2022_11_28/Internals/Secrets/SecretsService.swift
@@ -2,8 +2,8 @@ import Combine
 import Foundation
 
 internal protocol SecretsService {
-  func store(_ entry: SecretsServiceEntry, value: String) -> AnyPublisher<Void, StoreError>
-  func retrieve(_ entry: SecretsServiceEntry) -> AnyPublisher<String, RetrieveError>
+  func store(_ entry: SecretsServiceEntry, value: String) async -> Result<Void, StoreError>
+  func retrieve(_ entry: SecretsServiceEntry) async -> Result<String, RetrieveError>
 }
 
 internal enum SecretsServiceEntry {
