+++ Packages/Sources/Core/Octokit/Live/api_2022_11_28/Octokit+api_2022_11_28.swift
@@ -1,25 +1,25 @@
 import Octokit
 import Foundation
 
-internal extension Octokit {
-  static func api_2022_11_28() -> Self {
-    let config: Config = .api_2022_11_28()
-    let dispatchQueue: DispatchQueue = .global()
-    let networkClient: NetworkClient = URLSessionNetworkClient(deserializer: GithubApiJsonDeserializer())
-    let secretsService: SecretsService = KeychainSecretsService(dispatchQueue: .global())
-    
-    return Self(
-      loginService: {
-        LoginService.live(secretsService: secretsService)
-      },
-      userService: {
-        UserService.live(
-          config: config,
-          dispatchQueue: dispatchQueue,
-          networkClient: networkClient,
-          secretsService: secretsService
-        )
-      }
-    )
-  }
-}
+//internal extension Octokit {
+//  static func api_2022_11_28() -> Self {
+//    let config: Config = .api_2022_11_28()
+//    let dispatchQueue: DispatchQueue = .global()
+//    let networkClient: NetworkClient = URLSessionNetworkClient(deserializer: GithubApiJsonDeserializer())
+//    let secretsService: SecretsService = KeychainSecretsService(dispatchQueue: .global())
+//    
+//    return Self(
+//      loginService: {
+//        LoginService.live(secretsService: secretsService)
+//      },
+//      userService: {
+//        UserService.live(
+//          config: config,
+//          dispatchQueue: dispatchQueue,
+//          networkClient: networkClient,
+//          secretsService: secretsService
+//        )
+//      }
+//    )
+//  }
+//}
