+++ Packages/Sources/Core/Octokit/Live/api_2022_11_28/Octokit+init.swift
@@ -1,9 +1,9 @@
 import Octokit
 
-public extension Octokit {
-  init() {
-    self = .api_2022_11_28()
-  }
-  
-  static let shared: Self = .api_2022_11_28()
-}
+//public extension Octokit {
+////  init() {
+////    self = .api_2022_11_28()
+////  }
+////
+////  static let shared: Self = .api_2022_11_28()
+//}
